StonkLSTM/requirements.txt
matplotlib.pyplot
pandas
numpy
keras
sklearn
yfinance